# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Blog API
  version: 0.3.0
  description: |
    个人博客 HTTP API 契约。阶段二已提供管理员认证、文章编辑、
    分类标签和 OSS 媒体闭环；公开内容接口将在阶段三继续扩展。
servers:
  - url: /api/v1
    description: 当前站点
...
paths:
...
  /admin/categories:
    get:
      operationId: listAdminCategories
      summary: 查询全部分类
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 分类列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaxonomyItem"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createCategory
      summary: 新建分类
      tags: [Admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxonomyRequest"
      responses:
        "201":
          description: 分类已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxonomyItem"
        default:
          $ref: "#/components/responses/Problem"
  /admin/categories/{id}:
    parameters:
      - $ref: "#/components/parameters/TaxonomyId"
    put:
      operationId: updateCategory
      summary: 更新分类
      tags: [Admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxonomyRequest"
      responses:
        "200":
          description: 分类已更新
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxonomyItem"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteCategory
      summary: 删除分类
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "204":
          description: 分类已删除
        default:
          $ref: "#/components/responses/Problem"
  /admin/tags:
    get:
      operationId: listAdminTags
      summary: 查询全部标签
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 标签列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaxonomyItem"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createTag
      summary: 新建标签
      tags: [Admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxonomyRequest"
      responses:
        "201":
          description: 标签已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxonomyItem"
        default:
          $ref: "#/components/responses/Problem"
  /admin/tags/{id}:
    parameters:
      - $ref: "#/components/parameters/TaxonomyId"
    put:
      operationId: updateTag
      summary: 更新标签
      tags: [Admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxonomyRequest"
      responses:
        "200":
          description: 标签已更新
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxonomyItem"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteTag
      summary: 删除标签
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "204":
          description: 标签已删除
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TaxonomyId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    Problem:
      description: RFC 9457 问题详情
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
  schemas:
...
    TaxonomyOption:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
...
    TaxonomyRequest:
      type: object
      additionalProperties: false
      required: [name, slug, sortOrder, visible]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        slug:
          type: string
          minLength: 1
          maxLength: 80
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        description:
          type:
            - string
            - "null"
          maxLength: 300
        sortOrder:
          type: integer
          minimum: 0
        visible:
          type: boolean
    TaxonomyItem:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/TaxonomyOption"
        - type: object
          required:
            - sortOrder
            - visible
            - articleCount
            - createdAt
            - updatedAt
          properties:
            description:
              type:
                - string
                - "null"
            sortOrder:
              type: integer
            visible:
              type: boolean
            articleCount:
              type: integer
              format: int64
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    ProblemDetail:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        traceId:
          type: string
...